PC: Add suport for floating point values. Arduino: Embed some code to header file, use expander for controlling the backlight, change time management system, clean code, add compatibility with floating point temperatures and humidties.

+++ src/pl/karol202/weather/record/ForecastErrorRecord.java
@@ -2,7 +2,7 @@ package pl.karol202.weather.record;
 
 public class ForecastErrorRecord extends Record
 {
-	public ForecastErrorRecord(int timeInSeconds, int temperature, int humidity)
+	public ForecastErrorRecord(int timeInSeconds, float temperature, float humidity)
 	{
 		super(timeInSeconds, temperature, humidity);
 	}
